backend/requirements.txt
fastapi
uvicorn
httpx
python-dotenv
twilio
websockets
python-multipart
